Remove duplicate Album and ExternalIDs fields from FullTrack
They're both already present in SimpleTrack, and in the case of ExternalIDs,
in a more convenient form.  By putting the duplicate in FullTrack,
SimpleTrack is rendered invalid/incomplete (since the json tags on the
embedded fields are shadowed), and also use a less useful type
(map[string]string, rather than TrackExternalIDs).